requirements.txt
Flask==3.0.0
mysql-connector-python==8.1.0
opencv-python==4.7.0.72
face-recognition==1.3.0
numpy==1.26.0
flask-cors==4.0.0
openpyxl==3.1.2
werkzeug==3.0.0